+++ src/onegov.ticket/onegov/ticket/upgrade.py
@@ -10,7 +10,9 @@ from sqlalchemy import Column, Text
 @upgrade_task('Add handler_id to ticket', always_run=True)
 def add_handler_id_to_ticket(context):
 
-    if not context.has_column('tickets', 'handler_id'):
+    if context.has_column('tickets', 'handler_id'):
+        return False
+    else:
         context.operations.add_column(
             'tickets', Column(
                 'handler_id', Text, nullable=True, unique=True, index=True